Rename labwareSchemaShims.test.ts -> positionMath.test.ts
The file under test got renamed a while ago, but it looks like this test file got forgotten.
